chore: remove resources added to solve the problem
The problem was in fact in the main.tf of the cluster deployment! We need to be sure that Argo CD is deployed after all the cluster components. In our case Grafana was being deployed at the same time that the new Argo CD was and because of that Terraform received an error saying that the project was non existent.